# Sweep linestring intersection