Benchmark change:

```
benchmark                                               old ns/op      new ns/op     delta
BenchmarkIntersectsLineStringWithLineString/n=10        1051           7877          +649.48%
BenchmarkIntersectsLineStringWithLineString/n=100       108504         88597         -18.35%
BenchmarkIntersectsLineStringWithLineString/n=1000      12071435       937373        -92.23%
BenchmarkIntersectsLineStringWithLineString/n=10000     1103327828     11241364      -98.98%
```

The breakeven-point seems to be around n=~100. However, that would only apply to the particular input case from the benchmark. The main thing that the benchmark confirms is that the complexity class has improved.