@peterstace 2019-12-29
There are some remaining memory allocation overheads associated with interface usage (when using `line.Intersects(other)` and also when using the `heap` data structure). I'll tackle those in a second pass. We _should_ be able to make this a zero-allocation operation.